Changing file structure
Based on feedback on Moodle forum
