+++ index.html
@@ -12,7 +12,7 @@
 <body>
     <header class="col-12 position-fixed top-0">
         <nav class="navbar navbar-expand-md">
-            <div class="container">
+            <div class="container header">
                 <a href="index.html" class="navbar-brand"><i class="fa-solid fa-hand-holding-heart me-2"></i>SocialClub</a>
                 <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#nav" aria-expanded="false" aria-controls="#nav">
                     <span class="navbar-toggler-icon"></span>
@@ -35,8 +35,8 @@
     </header>
     <main>
         <div class="container pt-5">
-            <div class="row pt-4">
-                <div class="col-10 col-xl-6 col-lg-6 col-md-6 col-sm-6">
+            <div class="row pt-4 new-post-search">
+                <div class="col-xl-6 col-lg-6 col-md-6 col-sm-6 mt-2">
                     <button class="btn btn-primary" type="button" data-bs-toggle="collapse" data-bs-target="#new-post" aria-expanded="false" aria-controls="new-post">New post +</button>
                     <div class="collapse" id="new-post">
                         <div class="container">
@@ -56,7 +56,7 @@
                         </div>
                     </div>
                 </div>
-                <div class="col-2 col-xl-6 col-lg-6 col-md-6 col-sm-6 d-flex justify-content-end">
+                <div class="col-xl-6 col-lg-6 col-md-6 col-sm-6 d-flex justify-content-end mt-2">
                     <form class="d-flex align-items-center search-form">
                         <input type="text" id="search" name="search" placeholder="Search" class="form-control me-2">
                         <label for="search"><button class="btn btn-primary btn-search hidden-mobile"><i class="fa-solid fa-magnifying-glass"></i></button></label>
