+++ js/functions/logout.js
@@ -1,0 +1,7 @@
+import { logOut } from "../components/variables.js";
+
+export function setLogOut() {
+    logOut.addEventListener("click", (event) => {
+      localStorage.clear();
+    });
+}
