+++ login.html
@@ -19,9 +19,15 @@
                 </button>
                 <div class="collapse navbar-collapse justify-content-end" id="nav">
                     <ul class="navbar-nav">
-                    <li class="nav-item">
-                        <a href="index.html" class="nav-link">Home</a>
-                    </li>
+                        <li class="nav-item">
+                            <a href="index.html" class="nav-link">Home</a>
+                        </li>
+                        <li class="nav-item">
+                            <a href="create-account.html" class="nav-link">New user</a>
+                        </li>
+                        <li class="nav-item">
+                            <a href="login.html" class="nav-link active">Log in</a>
+                        </li>
                     </ul>
                 </div>
             </div>
